requirements.txt
ccxt
dataclasses
ipython
numpy
pandas
pymongo
requests
empyrical
python-binance
websocket-client
msgpack
scikit-learn
mongoengine
